requirements.txt
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
